fix: renamed dashboard page to file page
