+++ Frontend/src/components/Files/DeleteConfirmationModal.jsx
